move code related to EDF+ annotations to separate module
