+++ edfio/__init__.py
@@ -1,4 +1,5 @@
-from edfio.edf import Edf, EdfAnnotation, read_edf
+from edfio.edf import Edf, read_edf
+from edfio.edf_annotations import EdfAnnotation
 from edfio.edf_header import AnonymizedDateError, Patient, Recording
 from edfio.edf_signal import EdfSignal
