+++ tests/test_edf_annotations.py
@@ -12,7 +12,12 @@ from edfio import (
     EdfSignal,
     read_edf,
 )
-from edfio.edf import _create_annotations_signal, _EdfAnnotationsDataRecord
+from edfio.edf_annotations import (
+    _create_annotations_signal,
+    _EdfAnnotationsDataRecord,
+    _encode_annotation_duration,
+    _encode_annotation_onset,
+)
 from tests import TEST_DATA_DIR
 
 MNE_TEST_FILE = TEST_DATA_DIR / "mne_test.edf"
@@ -25,6 +30,56 @@ MNE_TEST_ANNOTATIONS = (
     EdfAnnotation(onset=2.0, duration=None, text="type B"),
     EdfAnnotation(onset=2.5, duration=2.5, text="type A"),
 )
+
+
+@pytest.mark.parametrize(
+    ("onset", "expected"),
+    [
+        (0, "+0"),
+        (0.0, "+0"),
+        (0.1, "+0.1"),
+        (0.01, "+0.01"),
+        (0.001, "+0.001"),
+        (0.0001, "+0.0001"),
+        (0.00001, "+0.00001"),
+        (0.000001, "+0.000001"),
+        (0.0000001, "+0.0000001"),
+        (0.00000001, "+0.00000001"),
+        (0.00000000001, "+0.00000000001"),
+        (100000000000.0, "+100000000000"),
+        (-0.1, "-0.1"),
+        (-0.0000001, "-0.0000001"),
+        (-0.0000000001, "-0.0000000001"),
+        (-100000000000.0, "-100000000000"),
+    ],
+)
+def test_encode_annotation_onset(onset: float, expected: str):
+    assert _encode_annotation_onset(onset) == expected
+
+
+@pytest.mark.parametrize(
+    ("duration", "expected"),
+    [
+        (0, "0"),
+        (0.0, "0"),
+        (0.1, "0.1"),
+        (0.01, "0.01"),
+        (0.001, "0.001"),
+        (0.0001, "0.0001"),
+        (0.00001, "0.00001"),
+        (0.000001, "0.000001"),
+        (0.0000001, "0.0000001"),
+        (0.00000000001, "0.00000000001"),
+        (100000000000.0, "100000000000"),
+    ],
+)
+def test_encode_annotation_duration(duration: float, expected: str):
+    assert _encode_annotation_duration(duration) == expected
+
+
+def test_encode_annotation_duration_raises_error_for_negative_values():
+    with pytest.raises(ValueError, match="Annotation duration must be positive, is"):
+        _encode_annotation_duration(-1)
 
 
 def test_edf_annotations():
